Added Queue to Query OpenAQ only right before we need the data
The Queue to OpenAQ makes sure that the Queue to Influxdb isn't full of older data than they need to be.

Also decreased the whole update frequency from 45 to 30 minutes since we will be filtering out the old data, so we can get through old data much quicker. Because of this, also added a check to make sure one update is finished before more gets queued up.