Got conda scripts to work
Needed to bind mount another folder to include the conda script. The script currently only includes the "grib_copy" part of the command. We are looking to try to use ofelia in order to schedule this script to run on its own as trying to run a "docker compose" command from a script on windows does not seem to work.